# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cosco API Documentation",
        "description": "API documentation for Cosco application",
        "contact": {
            "email": "admin@cosco.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/addresses": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get all addresses for the authenticated user",
                "operationId": "33ccea30443418dbab9242aa67931614",
                "responses": {
                    "200": {
                        "description": "List of addresses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AddressResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Address"
                ],
                "summary": "Create a new address for the authenticated user",
                "operationId": "5549e34fec9817a9794302a767fc7fde",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAddressRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddressResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/addresses/{id}": {
            "put": {
                "tags": [
                    "Address"
                ],
                "summary": "Update an address for the authenticated user",
                "operationId": "2f078db19c8d7e5a374faa4b048e4c45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Address ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAddressRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Address updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddressResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/article-categories/{slug}": {
            "get": {
                "tags": [
                    "ArticleCategory"
                ],
                "summary": "Get a single article category",
                "operationId": "77ab6db014887146131f8444c38019b1",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article category details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleCategoryResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Get paginated articles (optional search by title or category)",
                "operationId": "11aa4cc6561a7b94c56d1201d62513ee",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in article title or category (name/title/slug)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Results page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles list wrapped in ApiResponse with pagination",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "عملیات با موفقیت انجام شد"
                                        },
                                        "data": {
                                            "properties": {
                                                "articles": {
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/ArticleResource"
                                                            }
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 120
                                                        },
                                                        "perPage": {
                                                            "type": "integer",
                                                            "example": 12
                                                        },
                                                        "currentPage": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "lastPage": {
                                                            "type": "integer",
                                                            "example": 10
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "errors": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{slug}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Get a single article",
                "operationId": "37d4774ebe4d73de4d890c37720fcd74",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Article slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/brands/{slug}": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Get brand details with products",
                "operationId": "638d0dbda488c6cc27fd281b5783ef8a",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Brand slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Brand details with products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BrandResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreAddressRequest": {
                "required": [
                    "country_id",
                    "province_id",
                    "city_id",
                    "postal_code",
                    "plaque",
                    "address",
                    "phone"
                ],
                "properties": {
                    "country_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "province_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "postal_code": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "plaque": {
                        "type": "string",
                        "example": "12"
                    },
                    "address": {
                        "type": "string",
                        "example": "خیابان انقلاب، پلاک 12"
                    },
                    "phone": {
                        "type": "string",
                        "example": "02112345678"
                    },
                    "is_default": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "UpdateAddressRequest": {
                "properties": {
                    "country_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "province_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 100
                    },
                    "postal_code": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "plaque": {
                        "type": "string",
                        "example": "12"
                    },
                    "address": {
                        "type": "string",
                        "example": "خیابان انقلاب، پلاک 12"
                    },
                    "phone": {
                        "type": "string",
                        "example": "02112345678"
                    },
                    "is_default": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "AddressResource": {
                "title": "Address Resource",
                "description": "User address resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "postal_code": {
                        "description": "Postal code",
                        "type": "string",
                        "example": "1234567890"
                    },
                    "plaque": {
                        "description": "Building plaque number",
                        "type": "string",
                        "example": "12"
                    },
                    "address": {
                        "description": "Full address",
                        "type": "string",
                        "example": "خیابان انقلاب، پلاک 12"
                    },
                    "phone": {
                        "description": "Phone number",
                        "type": "string",
                        "example": "02112345678"
                    },
                    "is_default": {
                        "description": "Is default address",
                        "type": "boolean",
                        "example": true
                    },
                    "province_title_fa": {
                        "description": "Province name in Persian",
                        "type": "string",
                        "example": "تهران"
                    },
                    "province_id": {
                        "description": "Province ID",
                        "type": "integer",
                        "example": 10
                    },
                    "city_title_fa": {
                        "description": "City name in Persian",
                        "type": "string",
                        "example": "تهران"
                    },
                    "city_id": {
                        "description": "City ID",
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
            "ArticleCategoryResource": {
                "title": "Article Category Resource",
                "description": "Article category resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "اخبار سایت"
                    },
                    "slug": {
                        "type": "string",
                        "example": "news"
                    },
                    "is_show": {
                        "type": "boolean",
                        "example": true
                    },
                    "description": {
                        "type": "string",
                        "example": "دسته اخبار سایت"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/articles/news.jpg"
                    },
                    "seo_title": {
                        "type": "string",
                        "example": "سئو اخبار سایت"
                    },
                    "seo_description": {
                        "type": "string",
                        "example": "توضیحات سئو اخبار سایت"
                    },
                    "seo_follow": {
                        "type": "boolean",
                        "example": true
                    },
                    "seo_index": {
                        "type": "boolean",
                        "example": true
                    },
                    "seo_canonical": {
                        "type": "string",
                        "example": "https://example.com/news"
                    },
                    "articles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ArticleResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-27T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-27T12:34:56Z"
                    }
                },
                "type": "object"
            },
            "ArticleResource": {
                "title": "Article Resource",
                "description": "Article resource representation",
                "properties": {
                    "category_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "title": {
                        "type": "string",
                        "example": "عنوان مقاله"
                    },
                    "slug": {
                        "type": "string",
                        "example": "article-title"
                    },
                    "excerpt": {
                        "type": "string",
                        "example": "خلاصه مقاله"
                    },
                    "body": {
                        "type": "string",
                        "example": "متن کامل مقاله"
                    },
                    "is_show": {
                        "type": "boolean",
                        "example": true
                    },
                    "image_url": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/articles/article.jpg"
                    },
                    "view_count": {
                        "type": "integer",
                        "example": 123
                    },
                    "is_future": {
                        "type": "boolean",
                        "example": false
                    },
                    "seo_title": {
                        "type": "string",
                        "example": "سئو مقاله"
                    },
                    "seo_description": {
                        "type": "string",
                        "example": "توضیحات سئو مقاله"
                    },
                    "seo_follow": {
                        "type": "boolean",
                        "example": true
                    },
                    "seo_index": {
                        "type": "boolean",
                        "example": true
                    },
                    "seo_canonical": {
                        "type": "string",
                        "example": "https://example.com/article"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-27T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-27T12:34:56Z"
                    }
                },
                "type": "object"
            },
            "BrandResource": {
                "title": "Brand Resource",
                "description": "Brand resource schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "اپل"
                    },
                    "slug": {
                        "type": "string",
                        "example": "apple"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/brands/apple.png"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}